Run lint. Minor refactoring
