+++ ftt/ui/model.py
